requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
nextcord
Pillow
opencv-python
numpy
selenium
webdriver-manager
gspread
oauth2client
